# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kee — Cognitive Software Asset Intelligence System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vault:
    get:
      operationId: listVaultEntries
      tags: [vault]
      summary: List all Master Vault entries
      parameters:
        - name: tag
          in: query
          required: false
          schema:
            type: string
        - name: projectId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Vault entries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VaultEntry"
    post:
      operationId: createVaultEntry
      tags: [vault]
      summary: Add a new entry to the Master Vault
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultEntryInput"
      responses:
        "201":
          description: Created vault entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultEntry"
  /vault/{id}:
    get:
      operationId: getVaultEntry
      tags: [vault]
      summary: Get a vault entry by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vault entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultEntry"
    patch:
      operationId: updateVaultEntry
      tags: [vault]
      summary: Update a vault entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultEntryUpdate"
      responses:
        "200":
          description: Updated vault entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VaultEntry"
    delete:
      operationId: deleteVaultEntry
      tags: [vault]
      summary: Delete a vault entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    VaultEntry:
      type: object
      required: [id, title, content, entryType, createdAt, updatedAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        entryType:
          type: string
          enum: [note, document, conversation, idea, decision]
        tags:
          type: array
          items:
            type: string
        linkedProjectId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    VaultEntryInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        entryType:
          type: string
          enum: [note, document, conversation, idea, decision]
          default: note
        tags:
          type: array
          items:
            type: string
        linkedProjectId:
          type: integer
    VaultEntryUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        entryType:
          type: string
          enum: [note, document, conversation, idea, decision]
        tags:
          type: array
          items:
            type: string
        linkedProjectId:
          type: ["integer", "null"]
...