# lib/api-zod/src/generated/types/vaultEntry.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kee — Cognitive Software Asset Intelligence System API
 * OpenAPI spec version: 0.1.0
 */
import type { VaultEntryEntryType } from './vaultEntryEntryType';

export interface VaultEntry {
  id: number;
  title: string;
  content: string;
  entryType: VaultEntryEntryType;
  tags?: string[];
  /** @nullable */
  linkedProjectId?: number | null;
  createdAt: Date;
  updatedAt: Date;
}